Lazily initialize TextEditor widgets
This is just a proof of concept since it can have unwanted side-effects
due to never setting isInitialized for partviews and others. Needs more
checking where createWidget is being used and wether those places need
to include an explicit setInitialized() call.

The sublime API change is required because the mainwindow must not add
the XMLGUIClient for a view until the actual xmlgui children have all
been initialized, otherwise the xmlgui merging does not happen and that
can break said clients in weird ways.

This allows to lazily create the Kate view since its xmlgui is being
merged once it has been created and thus the contextmenu code in kate
will find the context menu element from the xmlgui file. Without the
lazy merging the katepartui stuff wouldn't be merged until one switches
to another document and back and thus the kate code looking in the xmlgui
file for the default context menu won't be finding anything.