Move the refactored code from IndexedStringNG to IndexedString.
We also increase the ItemRepository version number, but do not yet
port the rest of the code. This will be done in small steps later on.

Note that we also bring back/keep some old methods from IndexedString,
such as str() which is now deprecated. But also we have hash(), a KUrl
constructor etc.

In the future, we should introduce a separate IndexedPath or similar
which provides more efficient path/url serialization. Currently the
KUrl methods are very expensive due to the QString <-> KUrl
conversions...