Port ProjectModel to properly use IndexedString instead of raw indices
This is how the IndexedString should be used and not by using the raw
indices. Eventually we should also think about proper cleanup of the
database when URLs are removed from the project model, i.e. we should
properly increase/decrease the refcount then and delete items if the
refcount becomes 0.